backend/requirement.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
av==16.1.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.10
coverage==7.13.2
ctranslate2==4.6.3
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
faiss-cpu==1.13.2
fastapi==0.128.0
faster-whisper==1.2.1
filelock==3.20.3
Flask==3.1.0
flatbuffers==25.12.19
fsspec==2026.1.0
groq==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.4
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.7
langchain-core==1.2.7
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.6
MarkupSafe==3.0.2
moviepy==2.2.1
mpmath==1.3.0
numpy==2.4.1
onnxruntime==1.23.2
openai==2.16.0
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
passlib==1.7.4
pillow==11.3.0
pluggy==1.6.0
proglog==0.1.12
protobuf==6.33.4
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyjokes==0.8.3
pymongo==4.16.0
pypdf==6.6.2
pypiwin32==223
pyreadline3==3.5.4
pytest==9.0.2
pytest-cov==7.0.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pyttsx3==2.98
pywin32==310
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.2
tqdm==4.67.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
Werkzeug==3.1.3
xxhash==3.6.0
zstandard==0.25.0
